add register route
