--- pt-server/config/passport.js
@@ -1,47 +1,0 @@
-const passport = require('passport'),
-      User = require('../models/userModel'),
-      config = require('./main'),
-      JwtStrategy = require('passport-jwt').Strategy,
-      ExtractJwt = require('passport-jwt').ExtractJwt,
-      LocalStrategy = require('passport-local');
-/*
-const localOptions = { usernameField: 'email' };
-
-// Setting up local login strategy
-const localLogin = new LocalStrategy(localOptions, function(email, password, done) {
-  User.findOne({ email: email }, function(err, user) {
-    if(err) { return done(err); }
-    if(!user) { return done(null, false, { error: 'Incorrect email.' }); }
-
-    user.comparePassword(password, function(err, isMatch) {
-      if (err) { return done(err); }
-      if (!isMatch) { return done(null, false, { error: 'Incorrect password.' }); }
-
-      return done(null, user);
-    });
-  });
-});
-*/
-const jwtOptions = {
-  // Telling Passport to check authorization headers for JWT
-  jwtFromRequest: ExtractJwt.fromAuthHeaderAsBearerToken(),
-  // Telling Passport where to find the secret
-  secretOrKey: config.secret
-};
-
-
-// Setting up JWT login strategy
-const jwtLogin = new JwtStrategy(jwtOptions, function(payload, done) {
-  User.findById(payload._id, function(err, user) {
-    if (err) { return done(err, false); }
-
-    if (user) {
-      done(null, user);
-    } else {
-      done(null, false);
-    }
-  });
-});
-
-passport.use(jwtLogin);
-//passport.use(localLogin);
