+++ pt-server/config/passportJWT.js
@@ -1,0 +1,27 @@
+const passport = require('passport');
+const User = require('../models/userModel');
+const config = require('./main');
+const JwtStrategy = require('passport-jwt').Strategy;
+const ExtractJwt = require('passport-jwt').ExtractJwt;
+
+const jwtOptions = {
+  // Telling Passport where to extract the token from
+  jwtFromRequest: ExtractJwt.fromAuthHeaderAsBearerToken(),
+  // Telling Passport where to find the secret
+  secretOrKey: config.secret
+};
+
+// Setting up JWT login strategy
+const jwtLogin = new JwtStrategy(jwtOptions, function(payload, done) {
+  User.findById(payload._id, function(err, user) {
+    if (err) { return done(err, false); }
+
+    if (user) {
+      done(null, user);
+    } else {
+      done(null, false);
+    }
+  });
+});
+
+passport.use(jwtLogin);
